I am working on fixing the eevdf compiler errors.

+++ src/system/kernel/scheduler/HashMap.h
@@ -63,7 +63,7 @@ public:
             fBucketCount = newBucketCount;
         }
 
-        uint32 hash = (uint32)key % fBucketCount;
+        uint32 hash = (uintptr_t)key % fBucketCount;
         while (fBuckets[hash].key != 0) {
             hash = (hash + 1) % fBucketCount;
         }
